Can you change the script to accept an argument, so I can pass the file on the command line when invoking the script?
Sure! Here is the updated script that accepts a file path as a
command-line argument: (see commit)

To run the script with the file path as an argument, use the following
command in your terminal:

    node convert-svg-to-apng.js /Users/bosch/chessfun/chessboard.svg

This will pass the SVG file path to the script, which will then
process the file accordingly.