Prompt: I got the following error:
    $ node svg2apng.js SVG_animated_loading_icon.svg
    /Users/bosch/chessfun/node_modules/puppeteer-core/lib/cjs/puppeteer/cdp/Frame.js:183
    ? new Error(${response.errorText} at ${url}) ^

    Error: net::ERR_INVALID_URL at file://SVG_animated_loading_icon.svg
    at navigate
    (/Users/bosch/chessfun/node_modules/puppeteer-core/lib/cjs/puppeteer/cdp/Frame.js:183:27)
    at async Deferred.race
    (/Users/bosch/chessfun/node_modules/puppeteer-core/lib/cjs/puppeteer/util/Deferred.js:36:20)
    at async CdpFrame.goto
    (/Users/bosch/chessfun/node_modules/puppeteer-core/lib/cjs/puppeteer/cdp/Frame.js:149:25)
    at async CdpPage.goto
    (/Users/bosch/chessfun/node_modules/puppeteer-core/lib/cjs/puppeteer/api/Page.js:574:20)
    at async /Users/bosch/chessfun/svg2apng.js:15:3

    Node.js v23.7.0

The error indicates that the file URL is not correctly formed. You
need to provide the absolute path to the SVG file. Here is the
updated script to ensure the file path is correctly formatted:

To run the script with the file path as an argument, use the following
command in your terminal:

    node convert-svg-to-apng.js
    /Users/bosch/chessfun/SVG_animated_loading_icon.svg

This should resolve the `ERR_INVALID_URL` error by ensuring the
file path is correctly formatted.