Prompt: Now I got the following error:
    ./svg2apng SVG_animated_loading_icon.svg error: too many
    arguments. Expected 0 arguments but got 1.

The error occurs because commander expects options to be defined
explicitly. To handle positional arguments correctly, you need to
define them explicitly in the commander configuration.

Here's how you can modify your script to handle the SVG file path
as a positional argument: (see commit)

This modification ensures that the SVG file path is treated as a
required positional argument.